requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
google-cloud-aiplatform>=1.38.0
google-cloud-firestore>=2.14.0
google-cloud-discoveryengine>=0.11.8
pydantic>=2.6.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
